Fixing IOS-35, where feedback window looks strange.
In this bug, the survey window is presented, then the ratings flow causes the feedback window to be shown. The feedback window is unable to find the parent window of the presenting view controller, and so the positioning and sizing is messed up. This fix will walk the windows in the main screen if it can't find the right window, and look for the key window on that screen, or the key window on any screen if that fails.