Fix for IOS-59, in which the backend's working/not working logic was incorrect.
This fix clarifies the state consideration of whether the backend starts working.
The three pieces of controlling state are:

- is the network available?
- is the API key set?
- is the app entering the background or quitting?

In addition, the backend may attempt to send feedback immediately. This is no big
deal, because the user at that point is actively using the app. If the feedback fails
at that point, it will be retried when either the API key or network status changes,
so we're good there.

In general we want to be a little aggressive about starting the task queue, and the
above 3 conditions are helpful for either being aggressive about stopping the task
queue, or restarting it when the network or API key changes.