Fixing APPTENTIVE-497 Getting surveys returns 404 on no surveys
On getting /surveys/active, we get a 404 response if there is no survey for the app. This will prevent the configuration from being updated properly some of the time.

So, our temporary fix here, until we move to the new surveys API, is allow that task to fail silently. This should be fine.