Fixing bug where app is backgrounded after showing ratings dialog.
WHen the app is backgrounded while the "Do you love?" dialog is shown, then foregrounded, the viewController in ATAppRatingFlow may be set to nil due to calls to showRatingFlowFromViewControllerIfConditionsAreMet: by the app when the app enters the foreground.

So, here we will check to see if we can grab the root view controller, and if so, just use that to do our work.