Candidate fix for crashes related to a bad migration and leftover SQLite WALs.
The current theory being that while we migrated a store, we left over the WAL and SHM SQLite files, which resulted in database corruption.

Since the corruption involves missing columns in SQLite, we can try to detect that and wipe the database.

We also want to ensure that messages sent by the user are restored as such, and do so in ATGetMessagesTask.