Set ApptentiveResources deployment target back to 6.1
Setting the deployment target to 7 and above will compile the asset
catalog into a .car file, which must be in the main bundle for
`UIImage`'s `-imageNamed:` method to find it.

Once we drop iOS 7 support, we can use the new
`-imageNamed:inBundle:compatibleWithTraitCollection:`
method to specify the ApptentiveResources bundle.