Some work on tests for targeting fields and types.
Also introduces OCMock into the project, which exposed a problem in the way we were building Unit Tests (lots of duplicate objects in the linked binary).

Going forward, we shouldn't add new classes from ApptentiveConnect to the unit tests, and instead just pull them in from the linked libApptentiveConnect.

A couple of the unit tests are broken as it's not clear what the backward compatibility story here is.

IOSOSX-344, IOSOSX-341, IOSOSX-338