Avoid unintentional recursion/co-routine
In certain cases tasks (e.g. messages with missing nonces) will set their `finished` flag synchronously in their start method (rather than in response to a URL loading system callback), which triggers the KVO observation method, which calls the start method on another event. The result is that the calls are made recursively, rather than iteratively. This results in quickly running up an anomalous number of stack frames in instances where there are a large number of queued events. By running the call to -start asynchronously, this commit allows the KVO observation method to exit before the start method is next called.