Add strong reference from VC to interaction controller
The interaction controller instances were being dealloc'ed before their associated view controllers were dismissed, so we were losing the ability to close the interactions via a notification. This adds a strong reference to the interaction controller from the view controller so that the latter won't be deallocated until the former is dismissed.